Address PR feedback and separate out the force project reload functionality into a separate interceptor type.
